All impersonation steps refactored
