+++ Projects/RMAutomationUIFramework/src/test/java/org/fundacionjala/automation/scenario/steps/admin/impersonation/ImpersonationThenSteps.java
@@ -2,6 +2,10 @@ package org.fundacionjala.automation.scenario.steps.admin.impersonation;
 
 import java.util.List;
 
+import org.fundacionjala.automation.framework.pages.admin.emailserver.AddEmailServerPage;
+import org.fundacionjala.automation.framework.pages.admin.emailserver.EmailServerPage;
+import org.fundacionjala.automation.framework.pages.admin.impersonation.ImpersonationPage;
+import org.fundacionjala.automation.framework.pages.admin.login.LoginPage;
 import org.fundacionjala.automation.framework.pages.tablet.scheduler.CredentialsPage;
 import org.fundacionjala.automation.framework.utils.api.managers.ServiceAPIManager;
 import org.fundacionjala.automation.framework.utils.api.managers.SettingsAPIManager;
@@ -110,6 +114,51 @@ public class ImpersonationThenSteps {
 			if(impersonate == true) {	
 				ServiceAPIManager.putImpersonationRequest("http://172.20.208.84:4040/services", id, false);
 			}
+		}
+	}
+	
+	@Then("^Impersonation Option is disabled$")
+	public void Impersonation_Option_is_disabled() throws Throwable {
+		EmailServerPage emailServer = new EmailServerPage();
+		boolean impersonate = false;
+		List<Service> listServices;
+		listServices = ServiceAPIManager.getRequest("http://172.20.208.84:4040/services");
+		
+		for (Service service : listServices) {
+			impersonate = service.impersonate;
+		}
+		
+		ImpersonationPage impersonation = emailServer
+				.leftMenu
+				.clickOnImpersonationButton();
+		
+		boolean isSaveButtonPresent = impersonation.findSaveButton();
+		
+		Assert.assertFalse(isSaveButtonPresent);
+		Assert.assertFalse(impersonate);
+		
+		LoginPage login = new LoginPage();
+		
+		EmailServerPage server = login
+				.setUserName("SamuelSahonero")
+				.setPassword("Control*123")
+				.clickOnSigInButton()
+				.refreshPage()
+				.leftMenu
+				.clickOnEmailServerButton();
+		
+		boolean isEmailServerPresent = server.findEmailServer();
+		
+		if(isEmailServerPresent == false) {
+			
+			AddEmailServerPage addEmailServer = server
+					.clickOnAddButton();
+			
+			server = addEmailServer
+					.setDomainServer("roommanager.local")
+					.setUserName("Administrator")
+					.setPassword("Control*123")
+					.clickSaveButton();
 		}
 	}
 }
